+++ src/utils/mod.rs
@@ -1,4 +1,5 @@
 pub mod address_migration;
 pub mod keys;
+pub mod redis;
 pub mod secret_string;
 pub mod stream_helpers;
